requirements.txt
# Core dependencies
fastapi==0.95.0
uvicorn==0.21.1
python-dotenv==1.0.0
python-multipart==0.0.6
pydantic==1.10.7
requests==2.28.2
numpy==1.24.2
jinja2==3.1.2
httpx==0.23.3
aiofiles==23.1.0
itsdangerous==2.1.2

# Authentication
google-auth==2.17.1
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
firebase-admin==6.2.0

# Google Cloud
google-api-python-client==2.83.0
google-cloud-storage==2.8.0
google-cloud-firestore==2.11.0

# Video processing
opencv-python==4.7.0.72
ffmpeg-python==0.2.0
moviepy==1.0.3

# Image processing
Pillow==9.4.0
scikit-image==0.20.0

# AI and ML
scikit-learn==1.2.2
librosa==0.10.0.post2
tensorflow==2.12.0
runwayml==2.3.5
transformers==4.29.0
matplotlib==3.7.1
soundfile==0.12.1

# Audio processing
audioread==3.0.0
pydub==0.25.1

# Social media APIs
python-tiktok==0.1.4
facebook-sdk==3.1.0
instagrapi==1.17.3
google-api-python-client==2.83.0
youtube-dl==2021.12.17

# Scheduling and task management
APScheduler==3.10.1
celery==5.2.7
redis==4.5.4

# Visualization
plotly==5.14.1
dash==2.9.3
dash-bootstrap-components==1.4.1

# Data handling and caching
pandas==2.0.0
cachetools==5.3.0
xxhash==3.2.0

# Development
pytest==7.3.0
black==23.3.0
flake8==6.0.0
isort==5.12.0
mypy==1.2.0 